attached to shared
